+++ .env
@@ -1,0 +1,2 @@
+REACT_APP_API_ID=973f4fa3
+REACT_APP_API_KEY=d99e219fc4b58e878b793779677dd4ee
